Added collapsible sections to the toolboxes and refactored operations for localStorage
Each tool that have settings in the toolbox can now be collapsed.
The state is saved and loaded from localStorage.
The refactoring of operations towards localStorage was a major improvment.
Now a default stateobject (LOCAL_STORAGE_PROPS) is used in each tool that requires a saved state.
This communicates really well what properties that are saved and can easily be extended with more properties.
The magic of the spread operator makes it really easy to merge the saved state into a runtime copy of LOCAL_STORAGE_PROPS.