PR Release 1.0.0-beta2 (#2)
* Updated OL to version 6.13.0

* Updated incorrect information in README.md

* Updated form slider grip icon

* Changed to GitHub icon as the HiddenAbout-tool icon

* Updated spelling mistake

* Updated color modifier classes to match the naming order of color variables

* Added missing alpha value to the default background color for Markers

* Added collapsible sections to the toolboxes and refactored operations for localStorage

Each tool that have settings in the toolbox can now be collapsed.
The state is saved and loaded from localStorage.
The refactoring of operations towards localStorage was a major improvment.
Now a default stateobject (LOCAL_STORAGE_PROPS) is used in each tool that requires a saved state.
This communicates really well what properties that are saved and can easily be extended with more properties.
The magic of the spread operator makes it really easy to merge the saved state into a runtime copy of LOCAL_STORAGE_PROPS.

* Increased the min-width for the toolbox-container

This makes the collision-detection a bit more smooth when collapsing the toolbox sections.

* Added Table of contents and License information

* Added method to add settings from the individual tools

Thus decoupling responsibilities and removal of unnecessary settings when the corresponding tool is not used.

* Renamed constructor parameter to options instead of callbackObj

More and more tools require parameters that are not functions so i opted for a more general name.

* Implemented optional storage of data in localStorage for BookmarkTool

A constructor parameter (storeDataInLocalStorage) is used to enable localStorage support.

* Removed dead code

* Fixed jumping behavior when collapsing groups containing empty child elements

* Updated and added more information in the README-documentation

* Added function buttons in MyLocation marker popup

The user can now remove the marker even if the application is not using the LayerTool. It is also possible to copy the marker information.

* Spelling mistake

* Removed old file

* Improved documentation

* Refactored colors used in the light and dark theme

I felt that i were using to many colors making for a messy look-and-feel. Now trying to improve the meaning behind the color and just use a part of the color palette.

* Added more padding between the checkbox and the text

* Updated dependencies [axios, proj4]

* Refactored color system

Simplified the color-theme for the light and dark theme to have more colors in common. Also use the colors more consistent where the mid-color is to be considered the default color. Also improved the project documentation.

* Added module for easy access to GET parameters and added parameter to enable default context menu

Separated functionality for accessing get parameters in the URI. The module is called URIGet and is located in the helpers directory. Adding the ?debug=true parameter will now enable the built in context menu in the browser.

* Updated README.md documentation

* Updated dependencies [Parcel, ol, sass]

Updated Parcel-bundler from 1.12.5 to Parcel 2.4.1
Updated ol from 6.13.0 to 6.14.1
Replaced sass 1.49.9 with @parcel/transformer-sass 2.4.1

* Added variable for controlling border-radius on all components

* Bugfix for Magnifier tool

Re-assignment of constant variable. Weird that this has been working previously.

* Modified URIGet to return the retrieved value

The returned value can now be compared for any value not only boolean values.

* Added map zoom, lat, lon, rotation to localStorage

Added state management for base map properties, zoom, latitude, longitude and rotation. This is controlled from the map.js file.

* Added new tool that shows overview of map

* Code cleaning

* Changed logic for where the scaleline tool is initiated

* Updated dependencies [Parcel, Parcel-transformer-sass]

* Updated README.md with Node- and NPM version info

* Bugfix - Changed the css specificity for the Overview Map tool

* Moved the mouse wheel listener from the window to the toolbarElement

* Refactored function for export/import vectorlayer as geojson

* Structured files inside the helpers folder

* Changed type of listener function for the contextmenu

* Added support for import/export vector layer as KML

Vector layers can now be import and exported as both GeoJSON and KML format. The KML format also holds the style information that is applied to the drawn features.

* Bugfix for tooltip when changing layer name

* Standardized responses from modal windows

* Standardized modal content DOM

* Code formatting

* Modal window title update

* Restricted to only allow .geojson and .kml format in file dialog

* Added constructor method for olLayer, olSource and olFormat

* Renamed data container from attributes to properties for feature related items

* Updated export for contextmenu module

* Added parameters to disable function buttons in the layer tool

Default the buttons are enabled through the default object instance. The buttons can now be disabled through the use of constructor parameters.

* Added parameters to disable creation of map- and feature layers in the layer tool

* Added textbox to bookmark tool so that a name can be given at creation

* Code formatting

* Added default options instance to all tools that take a options-object as parameter

* Updated dependencies removed Axios

* Code formatting

* Refactored KeyboardZoom and KeyboardPan condition to use targetNotEditable

This removes the bug that the zoom- and arrow keys are blocked from textboxes/inputs that have focus.

* Updated README.md

* Updated Version 1.0.0-beta2